refactor: extract logic out of mod.rs/lib.rs into named files (scene-schema, frame_globals)
Module roots should control exposure only. Pure mechanical move, zero
behavior change (verified: cargo check clean, scene-schema tests pass):
- scene-schema/src/lib.rs: inline #[cfg(test)] mod tests → tests.rs.
- frame_globals/mod.rs: FrameGlobals + impls + error + helper → globals.rs;
  mod.rs keeps module decls + re-exports (public paths preserved).
